+++ demo/pages/examples/multi-select.ts
@@ -18,7 +18,7 @@ function updateCode() {
       label: 'HTML',
       lang: 'markup',
       code: `
-<asset-picker id="picker"></asset-picker>
+<sfx-asset-picker id="picker"></sfx-asset-picker>
 
 <script>
   picker.config = {